+++ docs/source/xlnstorch/index.rst
@@ -1,11 +1,11 @@
 xlnstorch
 =========
 
-xlnstorch is a PyTorch extension that provides a set of tools for
-simulating LNS arithmetic in machine learning models.
+xlnstorch is a PyTorch extension that provides a set of tools for simulating
+LNS arithmetic in machine learning models.
 
 .. toctree::
     :maxdepth: 1
 
-    XLNS in PyTorch <xlnstorch>
     LNS Tensors <lnstensors>
+    Operations <operations>
