+++ docs/source/xlnstorch/operations.rst
@@ -1,0 +1,23 @@
+.. currentmodule:: xlnstorch
+.. automodule:: xlnstorch
+
+.. _operations-doc:
+
+Operations
+==========
+
+.. autosummary::
+    :toctree: generated
+    :nosignatures:
+
+    implements
+    get_implementation
+    set_default_implementation
+    get_default_implementation_key
+    override_implementation
+    apply_lns_op
+    align_lnstensor_bases
+    format_lnstensor_operands
+
+    operators.implement_sbdb
+    operators.sbdb
