--- docs/source/xlnstorch/xlnstorch.rst
@@ -1,30 +1,0 @@
-xlnstorch
-=========
-.. automodule:: xlnstorch
-.. currentmodule:: xlnstorch
-
-LNSTensors
-----------
-
-.. _tensor-creation-ops:
-
-Creation Operations
-~~~~~~~~~~~~~~~~~~~
-.. autosummary::
-    :toctree: generated
-    :nosignatures:
-
-    lnstensor
-
-.. _tensor-custom-ops:
-
-Custom Operations
-~~~~~~~~~~~~~~~~~
-.. autosummary::
-    :toctree: generated
-    :nosignatures:
-
-    implements
-    set_default_implementation
-    override_impl
-    apply_lns_op
