+++ src/xlnstorch/dispatch_table.py
@@ -1,0 +1,208 @@
+import functools
+import contextlib
+from typing import Callable, Generator, Tuple
+
+# _HANDLED_FUNCTIONS is a dictionary that maps torch functions to their
+# corresponding implementations for LNSTensor. Each key is a torch function
+# and the value is a dictionary mapping implementation keys to a tuple of
+# the LNSTensor implementation and its internal computation function.
+_HANDLED_FUNCTIONS = {}
+# _DEFAULT_IMPLEMENTATIONS is a dictionary that maps torch functions to their
+# default implementation keys. This is used to determine which implementation
+# to use by default when a torch function is called on an LNSTensor.
+_DEFAULT_IMPLEMENTATIONS = {}
+
+def implements(
+        torch_function: Callable,
+        lns_operation: Callable,
+        key: str | None = None,
+        default: bool = False
+    ) -> Callable:
+    """
+    A decorator to register a custom implementation for a given torch function.
+
+    This allows functions to be mapped to specific handlers in the LNS context
+    and optionally set as the default implementation for that function.
+
+    Parameters
+    ----------
+    torch_function : Callable
+        The torch function to be overriden.
+    lns_operation : Callable
+        The function that implements the given LNS operation to tensors. This
+        should be the computation that is performed on the internal torch tensor
+        representations for the LNSTensor objects.
+    key : str, optional
+        A unique key to identify the implementation. If not provided, the 
+        function's name will be used by default.
+    default : bool, optional
+        If True, this implementation will be set as the default for the
+        specified torch function. Defaults to False.
+
+    Returns
+    -------
+    Callable
+        The decorator that registers the function as an implementation for
+        the specific torch function.
+    """
+    def decorator(func):
+        function_key = key or func.__name__
+        functools.update_wrapper(func, torch_function)
+
+        if torch_function not in _HANDLED_FUNCTIONS:
+            _HANDLED_FUNCTIONS[torch_function] = {}
+        _HANDLED_FUNCTIONS[torch_function][function_key] = (func, lns_operation)
+
+        if default:
+            _DEFAULT_IMPLEMENTATIONS[torch_function] = function_key
+
+        return func
+    return decorator
+
+def get_implementation(torch_function: Callable, impl_key: str) -> Tuple[Callable, Callable]:
+    """
+    Get the implementation tuple for a given torch function and implementation key.
+
+    Parameters
+    ----------
+    torch_function : Callable
+        The torch function for which to get the implementation.
+    impl_key : str
+        The key identifying the specific implementation.
+
+    Returns
+    -------
+    tuple[Callable, Callable]
+        A tuple containing (implementation_function, lns_operation).
+
+    Raises
+    ------
+    ValueError
+        If no implementations are registered for the given torch function.
+        If the specified implementation key is not registered for the torch function.
+    """
+    if torch_function not in _HANDLED_FUNCTIONS:
+        raise ValueError("No implementations registered for the given torch function.")
+
+    if impl_key not in _HANDLED_FUNCTIONS[torch_function]:
+        raise ValueError(f"Implementation '{impl_key}' is not registered for {torch_function}.")
+
+    return _HANDLED_FUNCTIONS[torch_function][impl_key]
+
+def set_default_implementation(torch_function: Callable, impl_key: str) -> None:
+    """
+    Set the default implementation for a given torch function.
+
+    Parameters
+    ----------
+    torch_function : Callable
+        The torch function for which to set the default implementation.
+    impl_key : str
+        The key identifying the implementation to be set as default.
+
+    Raises
+    ------
+    ValueError
+        If no implementations are registered for the given torch function.
+        If the specified implementation key is not registered for the torch function.
+    """
+    if torch_function not in _HANDLED_FUNCTIONS:
+        raise ValueError("No implementations registered for the given torch function.")
+
+    if impl_key not in _HANDLED_FUNCTIONS[torch_function]:
+        raise ValueError(f"Implementation '{impl_key}' is not registered for {torch_function}.")
+
+    _DEFAULT_IMPLEMENTATIONS[torch_function] = impl_key
+
+def get_default_implementation_key(torch_function: Callable) -> str:
+    """
+    Get the default implementation key for a given torch function.
+
+    Parameters
+    ----------
+    torch_function : Callable
+        The torch function for which to get the default implementation.
+
+    Returns
+    -------
+    str
+        The key identifying the default implementation.
+
+    Raises
+    ------
+    ValueError
+        If no implementations are registered for the given torch function.
+        If no default implementation is set for the torch function.
+    """
+    if torch_function not in _HANDLED_FUNCTIONS:
+        raise ValueError("No implementations registered for the given torch function.")
+
+    if torch_function not in _DEFAULT_IMPLEMENTATIONS:
+        raise ValueError(f"No default implementation set for {torch_function}.")
+
+    return _DEFAULT_IMPLEMENTATIONS[torch_function]
+
+@contextlib.contextmanager
+def override_implementation(torch_function: Callable, impl_key: str) -> Generator[None, None, None]:
+    """
+    Temporarily override the default implementation for a torch function within a context.
+    This allows for testing or using a different implementation without permanently changing
+    the default.
+
+    Parameters
+    ----------
+    torch_function : Callable
+        The torch function for which the implementation is to be temporarily overridden.
+    impl_key : str
+        The key identifying the new implementation to use as default.
+
+    Yields
+    ------
+    None
+        The function yields control back to the context block.
+
+    Examples
+    --------
+    >>> with override_implementation(torch.add, 'custom_add_impl'):
+    >>>     # Inside this block, torch.add will use 'custom_add_impl'
+    >>>     pass
+    """
+    original_default = _DEFAULT_IMPLEMENTATIONS.get(torch_function)
+    set_default_implementation(torch_function, impl_key)
+
+    try:
+        yield
+    finally:
+        set_default_implementation(torch_function, original_default)
+
+def apply_lns_op(torch_function: Callable, *args, **kwargs):
+    """
+    Performs the computation for the default LNS implementation of a given
+    torch function in the logarithmic domain. This function is used to
+    apply the LNS internal operation defined in ``HANDLED_FUNCTIONS``.
+
+    Parameters
+    ----------
+    torch_function : Callable
+        The torch function for which the LNS operation is to be applied.
+    *args : Any
+        Positional arguments to be passed to the LNS operation.
+    **kwargs : Any
+        Keyword arguments to be passed to the LNS operation.
+
+    Returns
+    -------
+    Any
+        The result of the LNS operation applied to the provided arguments.
+
+    Raises
+    ------
+    ValueError
+        If no implementations are registered for the given torch function.
+        If no default implementation is set for the torch function.
+        If the implementation key is not registered for the torch function.
+    """
+    impl_key = get_default_implementation_key(torch_function)
+    impl = get_implementation(torch_function, impl_key)
+
+    return impl[1](*args, **kwargs) # internal computation function
