+++ src/xlnstorch/operators/__init__.py
@@ -1,1 +1,9 @@
-from . import arithmetic_ops
+from .arithmetic_ops import (
+    implement_sbdb,
+    sbdb,
+)
+
+__all__ = [
+    "implement_sbdb",
+    "sbdb",
+]
